requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
qrcode==7.4.2
Pillow==10.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
